bubble up errors and move common podrunner function
